Increasing precision everywhere and removing dead code

+++ Corvette/Conventions.hh
@@ -11,7 +11,7 @@ namespace SGF {
 typedef unsigned int uint;
 typedef unsigned long ulong;
 typedef unsigned short ushort;
-typedef double MatrixElement;
+typedef long double MatrixElement;
 
 typedef uint occupancy_t;
 const uint Infinity=UINT_MAX; // The maximum possible integer is infinity
